+++ structurizr-core/src/com/structurizr/WorkspaceScope.java
@@ -1,0 +1,8 @@
+package com.structurizr;
+
+public enum WorkspaceScope {
+
+    Landscape,
+    SoftwareSystem
+
+}