+++ structurizr-core/src/com/structurizr/validation/WorkspaceScopeValidationException.java
@@ -1,0 +1,9 @@
+package com.structurizr.validation;
+
+public class WorkspaceScopeValidationException extends Exception {
+
+    public WorkspaceScopeValidationException(String message) {
+        super(message);
+    }
+
+}