+++ structurizr-core/src/com/structurizr/validation/WorkspaceScopeValidator.java
@@ -1,0 +1,9 @@
+package com.structurizr.validation;
+
+import com.structurizr.Workspace;
+
+public interface WorkspaceScopeValidator {
+
+    void validate(Workspace workspace) throws WorkspaceScopeValidationException;
+
+}