+++ structurizr-core/test/unit/com/structurizr/WorkspaceTests.java
@@ -73,4 +73,16 @@ public class WorkspaceTests {
         workspace.hydrate();
     }
 
+    @Test
+    void scope() {
+        Workspace workspace = new Workspace("Name", "Description");
+        assertNull(workspace.getScope()); // default scope is undefined
+
+        workspace.setScope(WorkspaceScope.SoftwareSystem);
+        assertEquals(WorkspaceScope.SoftwareSystem, workspace.getScope());
+
+        workspace.setScope(null);
+        assertNull(workspace.getScope());
+    }
+
 }